requirements.txt
customtkinter
pillow
google-genai
